Only allow owners to delete charts
This commit:

* Does not allow a user to delete a chart on the front-end unless they are
  the owner by disabling the button if they are not the owner.
* Does not allow a user to delete a chart on the back-end unless they are
  an owner by returning a 422 (unprocessible entity) if the user is not the
  chart owner.

+++ server/routes/charts.js
@@ -168,6 +168,16 @@ router
     try {
       const { chart_id } = req.params
       const { dataDb } = req.app.locals
+      const chart = await dataDb
+        .collection(collections.charts)
+        .findOne({ _id: new ObjectID(chart_id) })
+
+      if (chart.owner !== req.user) {
+        return res
+          .status(422)
+          .json({ message: 'Only the owner can delete a chart' })
+      }
+
       const deletedChart = await dataDb
         .collection(collections.charts)
         .deleteOne({ _id: ObjectID(chart_id) })
